# Conflicts:
#	Makefile
#	configure
#	include/mbgl/style/conversion/source.hpp
#	include/mbgl/style/sources/geojson_source.hpp
#	mbgl.gypi
#	package.json
#	platform/android/CHANGELOG.md
#	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/maps/MapView.java
#	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/maps/MapboxMap.java
#	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/maps/NativeMapView.java
#	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/maps/widgets/MyLocationView.java
#	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/style/layers/BackgroundLayer.java
#	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/style/layers/CircleLayer.java
#	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/style/layers/FillLayer.java
#	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/style/layers/Filter.java
#	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/style/layers/Layer.java
#	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/style/layers/LineLayer.java
#	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/style/layers/Property.java
#	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/style/layers/PropertyFactory.java
#	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/style/layers/RasterLayer.java
#	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/style/layers/SymbolLayer.java
#	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/style/sources/GeoJsonSource.java
#	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/style/sources/RasterSource.java
#	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/style/sources/Source.java
#	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/style/sources/VectorSource.java
#	platform/android/MapboxGLAndroidSDKTestApp/src/main/java/com/mapbox/mapboxsdk/testapp/activity/style/RuntimeStyleActivity.java
#	platform/android/MapboxGLAndroidSDKTestApp/src/main/res/layout/activity_runtime_style.xml
#	platform/android/MapboxGLAndroidSDKTestApp/src/main/res/menu/menu_runtime_style.xml
#	platform/android/MapboxGLAndroidSDKTestApp/src/main/res/raw/amsterdam.geojson
#	platform/android/MapboxGLAndroidSDKTestApp/src/main/res/values/strings.xml
#	platform/android/platform.gyp
#	platform/android/scripts/configure.sh
#	platform/android/scripts/generate-style-code.js
#	platform/android/src/jni.cpp
#	platform/android/src/style/android_conversion.hpp
#	platform/android/src/style/layers/background_layer.cpp
#	platform/android/src/style/layers/background_layer.hpp
#	platform/android/src/style/layers/circle_layer.cpp
#	platform/android/src/style/layers/circle_layer.hpp
#	platform/android/src/style/layers/fill_layer.cpp
#	platform/android/src/style/layers/fill_layer.hpp
#	platform/android/src/style/layers/layer.cpp
#	platform/android/src/style/layers/layer.hpp
#	platform/android/src/style/layers/layers.cpp
#	platform/android/src/style/layers/line_layer.cpp
#	platform/android/src/style/layers/line_layer.hpp
#	platform/android/src/style/layers/raster_layer.cpp
#	platform/android/src/style/layers/raster_layer.hpp
#	platform/android/src/style/layers/symbol_layer.cpp
#	platform/android/src/style/layers/symbol_layer.hpp
#	platform/android/src/style/sources/sources.cpp
#	platform/android/src/style/sources/sources.hpp
#	platform/default/mbgl/storage/offline_download.cpp
#	platform/ios/CHANGELOG.md
#	platform/ios/app/MBXViewController.m
#	platform/ios/ios.xcodeproj/project.pbxproj
#	platform/ios/scripts/configure.sh
#	platform/ios/uitest/OHHTTPStubs
#	platform/linux/scripts/configure.sh
#	platform/macos/macos.xcodeproj/project.pbxproj
#	platform/macos/scripts/configure.sh
#	platform/node/CHANGELOG.md
#	platform/qt/qmlapp/main.qml
#	platform/qt/scripts/configure.sh
#	platform/qt/src/qquickmapboxgl.cpp
#	platform/qt/src/qquickmapboxgl.hpp
#	platform/qt/src/qquickmapboxglrenderer.cpp
#	src/mbgl/annotation/line_annotation_impl.cpp
#	src/mbgl/renderer/painter_line.cpp
#	src/mbgl/style/layers/line_layer.cpp
#	src/mbgl/style/sources/geojson_source.cpp
#	src/mbgl/style/sources/geojson_source_impl.cpp
#	src/mbgl/style/sources/geojson_source_impl.hpp
#	src/mbgl/style/style.cpp
#	src/mbgl/tile/geojson_tile.cpp
#	src/mbgl/tile/geojson_tile.hpp
#	src/mbgl/util/geojson.cpp
#	test/test.gypi

+++ src/mbgl/geometry/collision_box_buffer.cpp
@@ -6,8 +6,8 @@
 
 namespace mbgl {
 
-size_t CollisionBoxVertexBuffer::add(int16_t x, int16_t y, float ox, float oy, float maxzoom, float placementZoom) {
-    const size_t idx = index();
+uint32_t CollisionBoxVertexBuffer::add(int16_t x, int16_t y, float ox, float oy, float maxzoom, float placementZoom) {
+    const uint32_t idx = index();
     void *data = addElement();
 
     int16_t *shorts = static_cast<int16_t *>(data);
@@ -17,7 +17,7 @@ size_t CollisionBoxVertexBuffer::add(int16_t x, int16_t y, float ox, float oy, f
     shorts[3] /* offset */ = ::round(oy);
 
     uint8_t *ubytes = static_cast<uint8_t *>(data);
-    // a_data1
+    // a_data
     ubytes[8] = maxzoom * 10;
     ubytes[9] = placementZoom * 10;
